+++ Incomes/Presentation/Component/ItemListYearSection.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2023 Hiromu Nakano. All rights reserved.
 //
 
-import Charts
 import SwiftData
 import SwiftUI
 
@@ -45,24 +44,7 @@ extension ItemListYearSection: View {
                 Text(title)
             })
             if isSubscribeOn {
-                Section("Balance") {
-                    Chart(items) {
-                        buildChartContent(date: $0.date,
-                                          value: $0.balance)
-                    }
-                    .frame(height: .componentL)
-                    .padding()
-                }
-                Section("Income and Outgo") {
-                    Chart(items) {
-                        buildChartContent(date: $0.date,
-                                          value: $0.income)
-                        buildChartContent(date: $0.date,
-                                          value: $0.outgo * -1)
-                    }
-                    .frame(height: .componentL)
-                    .padding()
-                }
+                ChartSections(items: items)
             } else {
                 Section {
                     Advertisement(type: .native(.medium))
@@ -83,22 +65,6 @@ extension ItemListYearSection: View {
                                 willDeleteItems = []
                             }
                         ])
-        }
-    }
-}
-
-private extension ItemListYearSection {
-    @ChartContentBuilder
-    func buildChartContent(date: Date, value: Decimal) -> some ChartContent {
-        if value != .zero {
-            BarMark(x: .value("Date", date),
-                    y: .value("Amount", value),
-                    stacking: .unstacked)
-                .foregroundStyle(value.isPlus ? Color.accentColor : Color.red)
-                .opacity(.medium)
-            RectangleMark(x: .value("Date", date),
-                          y: .value("Amount", value))
-                .foregroundStyle(value.isPlus ? Color.accentColor : Color.red)
         }
     }
 }
